Title: 세 수의 합과 평균 구하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240127/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.cpp
@@ -5,5 +5,5 @@ int main() {
     cin >> a >> b >> c;
     cout << fixed;
     cout.precision(0);
-    cout << a + b + c << endl << double(a + b + c)/3;
+    cout << a + b + c << endl << (a + b + c)/3;
 }
